+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New logger to send crash reports to Raygun.
 - New logger to send exceptions to Google (Universal) Analytics.
 - New logger to send logs to an Elasticsearch instance.
+- New logger to send logs to a Loki instance.
+- New logger to send logs to Grafana Cloud service.
 - New 'environment' context metadata for all external loggers.
 - Compatibility with WordPress 5.7.
 - New setting to override live console and local access privileges when in development or staging environments (thanks to [sebastienserre](https://github.com/sebastienserre) for the suggestion).